edit

+++ app/src/main/java/com/nbscollege/fitnessapp/ui/category/CategoryViewModel.kt
@@ -28,7 +28,7 @@ data class CategoryDetails(
     val exerciseCount: String = "",
     val subTitle: String = "",
     val backgroundImageResourceId: Int = 0,
-    val gif: Float = 0f,
+    val gif: Int = 0,
     val route: String = "",
 )
 /**
